Buffer Manager 1.0 working

+++ src/vulkan/CommandsVk.hpp
@@ -124,8 +124,8 @@ namespace boitatah::vk
 
     struct MapMemoryVk{
         VkDeviceMemory memory;
-        uint32_t offset;
-        uint32_t size;
+        uint64_t offset;
+        uint64_t size;
     };
     
     struct UnmapMemoryVk{
